renamed mapper.rb --> Mapper.rb
