+++ lib/bitstampede/Mapper.rb
